Move network request to separate file.
